Remove workspace.go, move stripCodeFences to exec.go
